Move backend manipulation out of Util
Seems to me like Util should just contain generic, totally
Fuse-independent utilities. The list of valid backends is most closely
associated with the configuration. The backend-materialization function
goes with the compiler driver.
